TASK 1: Set the width of the image globally to 100%. This will make all the images width equal to the width of its parent.
TASK 2: Set an id to the parent of the <b>first image( i.e. 'coding-ninjas-logo.png')</b> to <b>"logo"</b> and set the absolute width of the image to <b>210px</b>.

TASK 3: Change the width of the article element with <b>"blog content"</b> as id from <b>'1000px'</b> to <b>'90%'</b> of the viewport width.

TASK 4: Change the size of the text of element with id <b>'site-title'</b> to <b>'1.8rem'</b>.